spi - QueryInfo: change to return QueryResultRowIterator spi - QueryResultRow/QueryResultRowIterator: new jcr2spi - Update implementation accordingly (note jcr:score now is returned with property type Double) spi2dav - Minor changes to keep it compiling; full implementation of changed interface to be done
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@497844 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/spi/jcr2spi/src/main/java/org/apache/jackrabbit/jcr2spi/query/ScoreNodeIterator.java
@@ -33,5 +33,5 @@ public interface ScoreNodeIterator extends NodeIterator {
      * @return the score of the node returned by {@link #nextNode()}.
      * @throws java.util.NoSuchElementException if there is no next node.
      */
-    public float getScore();
+    public double getScore();
 }
